feat: 扩展插件接口并增强后台输入与 OpenCV 能力
新增 OpenCV 模块，提供模板管理、多模板匹配、尺度匹配、特征/边缘/形状匹配和文件预处理管线。

扩展 COM、C++ 与 SWIG/Python 接口，导出 Cv*、GetCursorShape、GetColorNum、GetPicSize、LayoutWindows、BindWindowEx 等能力，并调整窗口句柄为 LONG_PTR。

重构 DX 后台输入注入链路，加入 InputHookClient、DirectInput/RawInput 转发、DX 键盘鼠标实现和鼠标形状采集。

修复 GDI/WGC/DXGI 截图资源释放与错误处理，整理 AStar、Pipe、WinApi、字库解析、图像匹配和线程池实现。

+++ tools/EasyCom.cpp
@@ -183,7 +183,7 @@ int setupA(const char *path) {
     // }
     strcpy(dllpathA, path);
     wchar_t *m_wchar;
-    int nlen = strlen(dllpathA);
+    int nlen = static_cast<int>(strlen(dllpathA));
     int len = MultiByteToWideChar(CP_ACP, 0, dllpathA, nlen, NULL, 0);
     m_wchar = new wchar_t[len + 1];
     MultiByteToWideChar(CP_ACP, 0, dllpathA, nlen, m_wchar, len);
@@ -200,7 +200,7 @@ int setupW(const wchar_t *path) {
     // if (!fin) {
     //	return 0;//file not exist
     // }
-    int nlen = wcslen(path);
+    int nlen = static_cast<int>(wcslen(path));
     wcscpy(dllpathW, path);
     char *m_char;
     int len = WideCharToMultiByte(CP_ACP, 0, path, nlen, NULL, 0, NULL, NULL);
